// src/app/data-access/models/device.model.ts
export interface DeviceDetail {
  deviceId: string;
  manufDeviceId: string;
  deviceName: string;
  deviceDesc: string;
  devicePac: string;
  clientDeviceCategId: string;
  clientDeviceCategName: string;
  networkId: string;
  networkName: string;
  manufacturerId: string;
  manufacturerName: string;
  manufDeviceTypeId: string;
  manufDeviceTypeName: string;
  assetId: string;
  assetName: string;
  statusId: string;
  statusName: string;
  lastSeenDate: Date;
  lastSeenMsgId: string;
  certificate: string;
  companyId?: string;
}
// ...
export interface DeviceCategoryDirectory {
  deviceCategId: string;
  deviceCategName: string;
  companyName: string;
  companyId?: string;
  networkId: string;
  networkName: string;
}

export interface DeviceCategoryView {
  deviceCategName: string;
  deviceCategDesc: string;
  companyId?: string;
  companyName: string;
}

export interface DeviceFormData {
  companyId: string;
  userId: string;
  deviceAdd?: DeviceDetail[];
  deviceDel?: { deviceId: string }[];
}